+++ Source/ColonyManagerRedux.Managers/ManagerJobs/Settings/ManagerSettings_Hunting.cs
@@ -8,6 +8,9 @@ namespace ColonyManagerRedux.Managers;
 [HotSwappable]
 internal sealed class ManagerSettings_Hunting : ManagerSettings
 {
+    public bool DefaultSyncFilterAndAllowed = true;
+    public ManagerJob_Hunting.HuntingTargetResource DefaultTargetResource =
+        ManagerJob_Hunting.HuntingTargetResource.Meat;
     public bool DefaultAllowHumanLikeMeat;
     public bool DefaultAllowInsectMeat;
     public bool DefaultUnforbidCorpses = true;
@@ -20,10 +23,72 @@ internal sealed class ManagerSettings_Hunting : ManagerSettings
             rect.width,
             rect.height - Margin);
 
-        Widgets_Section.BeginSectionColumn(panelRect, "Hunting.Settings", out Vector2 position, out float width);
-        Widgets_Section.Section(ref position, width, DrawAllowWeirdMeat, "ColonyManagerRedux.ManagerSettings.DefaultThresholdSettings".Translate());
+        Widgets_Section.BeginSectionColumn(
+            panelRect, "Hunting.Settings", out Vector2 position, out float width);
+        Widgets_Section.Section(
+            ref position,
+            width,
+            DrawTargetResource,
+            "ColonyManagerRedux.Hunting.ManagerSettings.DefaultTargetResource".Translate());
+        Widgets_Section.Section(
+            ref position,
+            width,
+            DrawSyncFilterAndAllowed,
+            "ColonyManagerRedux.ManagerSettings.DefaultThresholdSettings".Translate());
+        Widgets_Section.Section(
+            ref position,
+            width,
+            DrawAllowWeirdMeat,
+            "ColonyManagerRedux.ManagerSettings.DefaultThresholdSettings".Translate());
         Widgets_Section.Section(ref position, width, DrawUnforbidCorpses);
         Widgets_Section.EndSectionColumn("Hunting.Settings", position);
+    }
+
+    public float DrawSyncFilterAndAllowed(Vector2 pos, float width)
+    {
+        var rowRect = new Rect(
+            pos.x,
+            pos.y,
+            width,
+            ListEntryHeight);
+
+        Utilities.DrawToggle(
+            rowRect,
+            "ColonyManagerRedux.SyncFilterAndAllowed".Translate(),
+            "ColonyManagerRedux.Hunting.SyncFilterAndAllowed.Tip".Translate(),
+            ref DefaultSyncFilterAndAllowed);
+
+        return ListEntryHeight;
+    }
+
+    public float DrawTargetResource(Vector2 pos, float width)
+    {
+        var targetResource =
+            (ManagerJob_Hunting.HuntingTargetResource[])
+            Enum.GetValues(typeof(ManagerJob_Hunting.HuntingTargetResource));
+
+        var cellWidth = width / targetResource.Length;
+
+        var cellRect = new Rect(
+            pos.x,
+            pos.y,
+            cellWidth,
+            ListEntryHeight);
+
+        foreach (var type in targetResource)
+        {
+            Utilities.DrawToggle(
+                cellRect,
+                $"ColonyManagerRedux.Hunting.TargetResource.{type}".Translate(),
+                $"ColonyManagerRedux.Hunting.TargetResource.{type}.Tip".Translate(),
+                DefaultTargetResource == type,
+                () => DefaultTargetResource = type,
+                () => { },
+                wrap: false);
+            cellRect.x += cellWidth;
+        }
+
+        return ListEntryHeight;
     }
 
     public float DrawAllowWeirdMeat(Vector2 pos, float width)
@@ -55,6 +120,9 @@ internal sealed class ManagerSettings_Hunting : ManagerSettings
     {
         base.ExposeData();
 
+        Scribe_Values.Look(ref DefaultSyncFilterAndAllowed, "defaultSyncFilterAndAllowed", true);
+        Scribe_Values.Look(ref DefaultTargetResource, "defaultTargetResource",
+            ManagerJob_Hunting.HuntingTargetResource.Meat);
         Scribe_Values.Look(ref DefaultAllowHumanLikeMeat, "defaultAllowHumanLikeMeat", false);
         Scribe_Values.Look(ref DefaultAllowInsectMeat, "defaultAllowInsectMeat", false);
         Scribe_Values.Look(ref DefaultUnforbidCorpses, "defaultUnforbidCorpses", true);